+++ src/dreamstate/core/observing/ActionsStore.ts
@@ -1,0 +1,5 @@
+/**
+ * Class for nested action values.
+ */
+export class ActionsStore {
+}