+++ src/dreamstate/core/observing/ComputedValue.test.ts
@@ -3,9 +3,9 @@ import { shouldObserversUpdate } from "@/dreamstate/core/observing/shouldObserve
 
 describe("Computed value class", () => {
   it("Should not be checked when comparing values before manager update", () => {
-    const firstLoadableInlineObj = { nested: Object.assign(new ComputedValue(), { a: 1 }) };
-    const secondLoadableInlineObj = { nested: Object.assign(new ComputedValue(), { b: 2 }) };
+    const firstComputedValue = { nested: Object.assign(new ComputedValue(), { a: 1 }) };
+    const secondComputedValue = { nested: Object.assign(new ComputedValue(), { b: 2 }) };
 
-    expect(shouldObserversUpdate(firstLoadableInlineObj, secondLoadableInlineObj)).toBeFalsy();
+    expect(shouldObserversUpdate(firstComputedValue, secondComputedValue)).toBeFalsy();
   });
 });